Refatoração em typescript
