+++ src/api/routes/index.ts
@@ -1,13 +1,14 @@
-const express = require('express')
+import express from 'express';
 const Router = express.Router();
 
 //rotas
-const bonds = require('../data/bonds');
-const courses = require('../data/courses');
-const grades = require('../data/grades');
-const homeworks = require('../data/homeworks');
-const members = require('../data/members');
-const news = require('../data/news');
+
+import courses from '../v1/courses';
+import grades from '../v1/grades';
+import homeworks from '../v1/homeworks';
+import members from '../v1/members';
+import news from '../v1/news';
+import bonds from '../v1/bonds';
 
 Router.post('/bonds', (req, res) =>  bonds(req,res));
 Router.post('/courses', (req,res) =>  courses(req, res));
@@ -16,4 +17,4 @@ Router.post('/members', (req,res) =>  members(req, res));
 Router.post('/homeworks', (req,res) =>  homeworks(req, res));
 Router.post('/news', (req,res) =>  news(req, res));
 
-module.exports = Router;
+export default Router;
