+++ src/api/v1/courses.ts
@@ -1,9 +1,12 @@
-module.exports = async function (req, res) {
-  const Sigaa = require("sigaa-api").Sigaa;
+import { Sigaa } from "sigaa-api";
+import { Request, Response } from "express";
+import isEmpty from "../../util/isEmpty";
+import findValue from "../../util/findValue";
+const sigaa = new Sigaa({
+  url: "https://sigaa.ifsc.edu.br",
+});
+export default async function (req:Request, res:Response) {
 
-  const sigaa = new Sigaa({
-    url: "https://sigaa.ifsc.edu.br",
-  });
   var bondsJSON = [];
 
   const username = req.body.username;
@@ -17,35 +20,7 @@ module.exports = async function (req, res) {
   var allBonds = [];
   allBonds.push(activeBonds, inactiveBonds);
 
-  function isEmpty(val) {
-    let typeOfVal = typeof val;
-    switch (typeOfVal) {
-      case 'object':
-        return (val.length == 0) || !Object.keys(val).length;
-        break;
-      case 'string':
-        let str = val.trim();
-        return str == '' || str == undefined;
-        break;
-      case 'number':
-        return val == '';
-        break;
-      default:
-        return val == '' || val == undefined;
-    }
-  };
-
-  function findValue(args, obj) {
-    for (let [key_arg, value_arg] of Object.entries(args)) {
-      for (let [key, value] of Object.entries(obj)) {
-        if (key_arg == key && value_arg == value) {
-          return obj;
-        }
-      }
-    }
-  }
-
-  function pushCourses(course) {
+  function pushCourses(course:any) {
     return {
       id: course.id,
       title: course.title,
@@ -56,7 +31,7 @@ module.exports = async function (req, res) {
   }
 
 
-  function pushBonds(bond) {
+  function pushBonds(bond:any) {
     return {
       program: bond.program,
       registration: bond.registration,
@@ -64,13 +39,13 @@ module.exports = async function (req, res) {
     }
   }
 
-  function courseHandler(course) {
+  function courseHandler(course:any) {
     coursesJSON.push(pushCourses(course));
   }
   for (const bonds of allBonds) {
     for (let i = 0; i < bonds.length; i++) {
-      coursesJSON = [];
-      const bond = bonds[i];
+      var coursesJSON:any = [];
+      const bond:any = bonds[i];
       if (!isEmpty(args) && !findValue(args, bond)) break; // se tiver argumentos e não for valido
       const courses = await bond.getCourses();
       for (const course of courses) {
@@ -80,8 +55,8 @@ module.exports = async function (req, res) {
       bondsJSON.push(pushBonds(bond));
     }
   }
-  res.json({
+  await account.logoff();
+  return res.json({
     bonds: bondsJSON
   });
-  await account.logoff();
 }
