+++ src/api/v1/homeworks.ts
@@ -1,11 +1,14 @@
-module.exports = async function (req, res) {
-    const Sigaa = require("sigaa-api").Sigaa;
+import { Sigaa, SigaaStudentBond, CourseStudent, StudentBond, BondType, Homework, SigaaHomework } from 'sigaa-api';
+import { Request, Response } from "express";
+import isEmpty from "../../util/isEmpty";
+import findValue from "../../util/findValue";
 
-    const sigaa = new Sigaa({
-        url: "https://sigaa.ifsc.edu.br",
-    });
-    var bondsJSON = [];
-    var coursesJSON = [];
+const sigaa = new Sigaa({
+  url: "https://sigaa.ifsc.edu.br",
+});
+export default async function (req:Request, res:Response) {
+    var bondsJSON:any = [];
+    var coursesJSON:any = [];
 
     const username = req.body.username;
     const password = req.body.password;
@@ -18,42 +21,9 @@ module.exports = async function (req, res) {
     var allBonds = [];
     allBonds.push(activeBonds, inactiveBonds);
 
-    function isEmpty (val) {
-        let typeOfVal = typeof val;
-        switch(typeOfVal){
-            case 'object':
-                return (val.length == 0) || !Object.keys(val).length;
-                break;
-            case 'string':
-                let str = val.trim();
-                return str == '' || str == undefined;
-                break;
-            case 'number':
-                return val == '';
-                break;
-            default:
-                return val == '' || val == undefined;
-        }
-    };
-    function findValue(args, obj) {
-        for (let [key_arg, value_arg] of Object.entries(args)) {
-            for (let [key, value] of Object.entries(obj)) {
-                if (key_arg == key && value_arg == value) {
-                    return obj;
-                }
-            }
-        }
-    }
-
-    async function pushHomeworks(homeworkList) {
+    async function pushHomeworks(homeworkList:any) {
         var homeworks = [];
         for (const homework of homeworkList) {
-            try {
-                var file = await homework.getAttachmentFile();
-            } catch (err) {
-                var file = [];
-            }
-
             var description = await homework.getDescription();
             var haveGrade = (await homework.getHaveGrade()) ? 'Vale nota' : 'Não vale nota'
             var startDate = homework.startDate;
@@ -61,7 +31,6 @@ module.exports = async function (req, res) {
 
             homeworks.push({
                 title: homework.title,
-                fileTitle: file.title,
                 description: description,
                 startDate: startDate,
                 endDate: endDate,
@@ -71,7 +40,7 @@ module.exports = async function (req, res) {
         return homeworks;
     }
 
-    function pushCourses(course, homeworksJSON) {
+    function pushCourses(course: CourseStudent, homeworksJSON:object) {
         return {
             id: course.id,
             title: course.title,
@@ -82,13 +51,13 @@ module.exports = async function (req, res) {
         }
 
     }
-    async function homeworkHandler(course) {
-        const homeworkList = await course.getHomeworks();
+    async function homeworkHandler(course: CourseStudent) {
+        const homeworkList:any = await course.getHomeworks();
         const homeworksJSON = await pushHomeworks(homeworkList);
         coursesJSON.push(pushCourses(course, homeworksJSON))
     }
 
-    function pushBonds(bond) {
+    function pushBonds(bond:any) {
         return {
             program: bond.program,
             registration: bond.registration,
@@ -98,7 +67,7 @@ module.exports = async function (req, res) {
     for (const bonds of allBonds) {
         for (let i = 0; i < bonds.length; i++) {
             coursesJSON = [];
-            const bond = bonds[i];
+            const bond:any = bonds[i];
             if (!isEmpty(args) && !findValue(args, bond)) break; // se tiver argumentos e não for valido
             else if (isEmpty(args)) { //verifica se existem argumentos
                 res.json({
@@ -114,8 +83,8 @@ module.exports = async function (req, res) {
             bondsJSON.push(pushBonds(bond));
         }
     }
-    res.json({
+    await account.logoff();
+    return res.json({
         bonds: bondsJSON
     });
-    await account.logoff();
 }
