+++ src/api/v1/news.ts
@@ -1,11 +1,14 @@
-module.exports = async function (req, res) {
-    const Sigaa = require("sigaa-api").Sigaa;
+import { Sigaa, CourseStudent, MemberList, SigaaNews, News } from 'sigaa-api';
+import { Request, Response } from "express";
+import isEmpty from "../../util/isEmpty";
+import findValue from "../../util/findValue";
 
-    const sigaa = new Sigaa({
-        url: "https://sigaa.ifsc.edu.br",
-    });
-    var bondsJSON = [];
-    var coursesJSON = [];
+const sigaa = new Sigaa({
+  url: "https://sigaa.ifsc.edu.br",
+});
+export default async function (req: Request, res: Response) {
+    var bondsJSON:any = [];
+    var coursesJSON:any = [];
 
     const username = req.body.username;
     const password = req.body.password;
@@ -17,34 +20,8 @@ module.exports = async function (req, res) {
 
     var allBonds = [];
     allBonds.push(activeBonds, inactiveBonds);
-    function isEmpty (val) {
-        let typeOfVal = typeof val;
-        switch(typeOfVal){
-            case 'object':
-                return (val.length == 0) || !Object.keys(val).length;
-                break;
-            case 'string':
-                let str = val.trim();
-                return str == '' || str == undefined;
-                break;
-            case 'number':
-                return val == '';
-                break;
-            default:
-                return val == '' || val == undefined;
-        }
-    };
-    function findValue(args, obj) {
-        for (let [key_arg, value_arg] of Object.entries(args)) {
-            for (let [key, value] of Object.entries(obj)) {
-                if (key_arg == key && value_arg == value) {
-                    return obj;
-                }
-            }
-        }
-    }
 
-    function pushCourses(course, newsJSON) {
+    function pushCourses(course: CourseStudent, newsJSON:any) {
         return {
             id: course.id,
             title: course.title,
@@ -55,7 +32,7 @@ module.exports = async function (req, res) {
         }
     }
 
-    async function pushNews(newsList) {
+    async function pushNews(newsList: News[]) {
         var newsJSON = [];
         for (const news of newsList) {
             newsJSON.push({
@@ -68,13 +45,13 @@ module.exports = async function (req, res) {
         return newsJSON;
     }
 
-    async function newsHandler(course) {
+    async function newsHandler(course: CourseStudent) {
         const newsList = await course.getNews();
         const newsJSON = await pushNews(newsList);
         coursesJSON.push(pushCourses(course, newsJSON))
     }
 
-    function pushBonds(bond) {
+    function pushBonds(bond:any) {
         return {
             program: bond.program,
             registration: bond.registration,
@@ -84,7 +61,7 @@ module.exports = async function (req, res) {
     for (const bonds of allBonds) {
         for (let i = 0; i < bonds.length; i++) {
             coursesJSON = [];
-            const bond = bonds[i];
+            const bond:any = bonds[i];
             if (!isEmpty(args) && !findValue(args, bond)) break; // se tiver argumentos e não for valido
             else if (isEmpty(args)) { //verifica se existem argumentos
                 res.json({
@@ -100,8 +77,8 @@ module.exports = async function (req, res) {
             bondsJSON.push(pushBonds(bond));
         }
     }
-    res.json({
+    await account.logoff();
+    return res.json({
         bonds: bondsJSON
     });
-    await account.logoff();
 }
